Update loggers that print results of queries to use trace.
We want more logging detail, without either being swamped in
data or slowing down the application. This way we can see the queries in debug, but only the results when we're using trace.

Because we can see the queries, we can run them again manually if we need to examine the data.