Add logging filter (#96)
* Add metadata filter for better metrics.

We want to collect some metrics for usage. These are currently lost in
the logging because everything is a POST request which doesn't capture
the body.

* Replace Filter with Action to generate logging metadata.

Filter applies to all routes. By using Action we can only target the
routes which we are interested in.

* Externalise configuration for metadata logging.

* Remove new line characters from log statements.

The log statements generated in GCP were a mess because each new line
character creates a new entry rather than bundling them together into a
single log.