1887 result caching (#97)
* Update formatting using scalafmt.

* Add caching of graphql responses.

* Finetune caching mechanism.

Graphql typically return Http 200 even when there are errors (unless
the query can't be parsed). We don't want to cache errors which might
resolve themselves. We check for errors in the response body before
passing it back.

* Update Dockerfile to log logging configuration on start-up.

* Fix unsafe access to option in GraghQLController.scala.

* Update GqlRequestMetadata to use String instead of Option[String].

This case class is only used to store logging metadata, so we should
compel the caller to provide a default.

* Update production.xml to use recommended configuration.

The [Play documentation](https://www.playframework.com/documentation/2.8
.x/SettingsLogger) provides guidance on how the logging framework should
 be set up for use in production environments.