Add literature sentences DB table.
Previously literature sentences were included in the `ot.literature`
table. It was no longer viable to save massive JSON blobs in Clickhouse,
so a new table `ot.sentences` was added. This table holds the sentence
metadata for a specific publication.

The sentence metadata is used to provide highlighting on the EPMC
evidence widget.

Relates to opentargets/platform#1793